Inline code blocks are now red on white
Also separated init script and inline styles in a .js and a .css file as
embedded resources. The contents is inlined in the HTML before
generating.

Also added custom.css file to the HEAD links. If there is one in the
inut folder it will get picked up and included.

Also fixed the markdown files to be UTF-8 to fix the non ASCII characters issues