Implemented speaker notes
Another Markdig win ! After a bit of fumbling I realized that FsReveal style speaker notes
are basically blockquotes with a ' instead of a >

It then was a matter of registering a derived parser and modifying the
rendering code for the QuoteBlock. Had to tweak the code to output <br/>
at the end of the lines because otherwise Reveal display everything in a
single line. This turned out to be trivial once I found out about
LineBreakInlineRenderer.RenderAsHardlineBreak.