Automated slide printing. Migrated to latest ASP.NET Core
Uses the great PuppeteerSharp library. This library handles the
downloading and remote control of Chromium. Works pretty fast too, and
in parallel !

Changed project to be a web project with the ASP.NET assemblies loaded
as a Framework reference.

Added a --log-level CLI option, defaults to Error to avoid spammy
messages in the console.

Lots of little aesthetic fixes to the console output.